@@ -21,7 +21,7 @@
 		"zombie": "^6.1.4"
 	},
 	"engines": {
-		"node": "4.4.3"
+		"node": "13.8.0"
 	},
 	"repository": {
 		"type": "git",
