@@ -6,8 +6,8 @@
 *       (if additional are added, keep them at the very end!)
 */
 
-var chaiHttp = require('chai-http');
-var chai = require('chai');
+import chaiHttp from 'chai-http';
+import chai from 'chai';
 var assert = chai.assert;
 var server = require('../server');
 
